RELAND "GraphicBufferSource: Use 64bit consumer usage"
This reverts commit 5099027f2c6034857ce41be5a30bbe8979cc2a9e.

Bug: 295498401
Bug: 185896428
Test: forrest run https://android-build.googleplex.com/builds/forrest/run/L55400000962546914

+++ media/libstagefright/omx/OmxGraphicBufferSource.cpp
@@ -85,7 +85,7 @@ status_t OmxGraphicBufferSource::configure(
         int32_t bufferCount,
         uint32_t frameWidth,
         uint32_t frameHeight,
-        uint32_t consumerUsage) {
+        uint64_t consumerUsage) {
     if (omxNode == NULL) {
         return BAD_VALUE;
     }
